==> risks-lens/09-risks-lens-main.Rmd <==
---
# risks-lens-main.Rmd
#
title: "Through the lens of the RISKS Forum"
author: "Daniel Moul"
date: "2020-11-21"
output:
  bookdown::gitbook:
    config:
      sharing: no
      toc:
        before: |
          <li><strong><a href="https://dmoul.github.io/risks-lens/the-lens.html">Risks through the lens<br>of the RISKS Forum</a></strong></li>
        after: |
          <br><br><li><strong><a href="https://dmoul.github.io/">By Daniel Moul</a></strong></li>

params:
  save_figures: FALSE
  # other knitr options are set in risk-analysis-setup.R
---

```{r child='09A-risks-lens.Rmd'}
```

\newpage

```{r child='09B-risks-lens.Rmd', eval=TRUE}
```

\newpage

```{r child='09C-risks-lens.Rmd'}
```
